move wg ip logic to server/generate package and add RPC
