+++ client/command/generate.go
@@ -42,7 +42,6 @@ import (
 	"github.com/bishopfox/sliver/protobuf/clientpb"
 	"github.com/bishopfox/sliver/protobuf/commonpb"
 	"github.com/bishopfox/sliver/protobuf/rpcpb"
-	"github.com/bishopfox/sliver/server/db"
 	"github.com/desertbit/grumble"
 )
 
@@ -84,7 +83,7 @@ var (
 )
 
 func generate(ctx *grumble.Context, rpc rpcpb.SliverRPCClient) {
-	config := parseCompileFlags(ctx)
+	config := parseCompileFlags(ctx, rpc)
 	if config == nil {
 		return
 	}
@@ -270,7 +269,7 @@ func generateStager(ctx *grumble.Context, rpc rpcpb.SliverRPCClient) {
 }
 
 // Shared function that extracts the compile flags from the grumble context
-func parseCompileFlags(ctx *grumble.Context) *clientpb.ImplantConfig {
+func parseCompileFlags(ctx *grumble.Context, rpc rpcpb.SliverRPCClient) *clientpb.ImplantConfig {
 	var name string
 	if ctx.Flags["name"] != nil {
 		name = strings.ToLower(ctx.Flags.String("name"))
@@ -372,14 +371,14 @@ func parseCompileFlags(ctx *grumble.Context) *clientpb.ImplantConfig {
 	}
 
 	var tunIP net.IP
-	var err error
 	if wg := ctx.Flags.String("wg"); wg != "" {
-		tunIP, err = GenerateUniqueIP()
+		uniqueWGIP, err := rpc.GenerateUniqueIP(context.Background(), &commonpb.Empty{})
+		tunIP = net.ParseIP(uniqueWGIP.IP)
 		if err != nil {
-
-			fmt.Printf(Warn + "Failed to generate unique ip for wg peer tun interface")
+			fmt.Println(Warn + "Failed to generate unique ip for wg peer tun interface")
 			return nil
 		}
+		fmt.Printf(Info+"Generated unique ip for wg peer tun interface: %s\n", tunIP.String())
 	}
 
 	config := &clientpb.ImplantConfig{
@@ -738,7 +737,7 @@ func newProfile(ctx *grumble.Context, rpc rpcpb.SliverRPCClient) {
 		fmt.Printf(Warn + "Invalid profile name\n")
 		return
 	}
-	config := parseCompileFlags(ctx)
+	config := parseCompileFlags(ctx, rpc)
 	if config == nil {
 		return
 	}
@@ -850,77 +849,4 @@ func displayCanaries(canaries []*clientpb.DNSCanary, burnedOnly bool) {
 			fmt.Printf("%s\n", line)
 		}
 	}
-}
-
-// GenerateUniqueIP generates and returns an available IP which can then
-// be assigned to a Wireguard interface
-func GenerateUniqueIP() (net.IP, error) {
-	peersTunIps, err := db.WGPeerIPs()
-	if err != nil {
-		fmt.Printf(Warn+"Failed to retrieve list WG Peers IPs %s", err)
-		return nil, err
-	}
-
-	// Use the 100.64.0.1/16 range for TUN ips.
-	// This range chosen due to Tailscale also using it (Cut down to /16 instead of /10)
-	// https://tailscale.com/kb/1015/100.x-addresses
-	addressPool, err := hosts("100.64.0.1/16")
-	if err != nil {
-		fmt.Printf(Warn+"Failed to generate host address pool for WG Peers IPs %s", err)
-		return nil, err
-	}
-
-	for _, address := range addressPool {
-		for _, peerTunIp := range peersTunIps {
-			if peerTunIp == address {
-				addressPool = remove(addressPool, []string{peerTunIp})
-				break
-			}
-		}
-	}
-
-	return net.ParseIP(addressPool[0]), nil
-}
-
-// Reserve use of 100.64.0.{0|1} addresses
-var reservedAddresses = []string{"100.64.0.0", "100.64.0.1"}
-
-func hosts(cidr string) ([]string, error) {
-	ip, ipnet, err := net.ParseCIDR(cidr)
-	if err != nil {
-		return nil, err
-	}
-
-	var ips []string
-	for ip := ip.Mask(ipnet.Mask); ipnet.Contains(ip); incrementIP(ip) {
-		ips = append(ips, ip.String())
-	}
-
-	ips = remove(ips, reservedAddresses)
-	return ips, nil
-}
-
-func incrementIP(ip net.IP) {
-	for j := len(ip) - 1; j >= 0; j-- {
-		ip[j]++
-		if ip[j] > 0 {
-			break
-		}
-	}
-}
-
-func remove(stringSlice []string, remove []string) []string {
-	var result []string
-	for _, v := range stringSlice {
-		shouldAppend := true
-		for _, value := range remove {
-			if v == value {
-				shouldAppend = false
-			}
-		}
-		if shouldAppend {
-			result = append(result, v)
-		}
-	}
-	return result
 }
